+++ stdlib/number/mod_test.go
@@ -1,6 +1,7 @@
 package numberfunc_test
 
 import (
+	"math/big"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
@@ -47,6 +48,24 @@ func TestModFunction(test *testing.T) {
 			Expected: function.RunResponse{
 				Result: resultData,
 				Error:  function.NewArgumentFuncError(1, "mod: divisor cannot be zero"),
+			},
+		},
+		"dividend-beyond-upper-limit": {
+			Request: function.RunRequest{
+				Arguments: function.NewArgumentsData([]attr.Value{types.NumberValue(func() *big.Float { f := new(big.Float); f.SetString("1e+310"); return f }()), types.Float64Value(10)}),
+			},
+			Expected: function.RunResponse{
+				Result: resultData,
+				Error:  function.NewArgumentFuncError(0, "mod: dividend is beyond the limits of float64"),
+			},
+		},
+		"divisor-beyond-upper-limit": {
+			Request: function.RunRequest{
+				Arguments: function.NewArgumentsData([]attr.Value{types.Float64Value(10), types.NumberValue(func() *big.Float { f := new(big.Float); f.SetString("1e+310"); return f }())}),
+			},
+			Expected: function.RunResponse{
+				Result: resultData,
+				Error:  function.NewArgumentFuncError(1, "mod: divisor is beyond the limits of float64"),
 			},
 		},
 	}